External DB credentials, install script
The database credentials, host address, and name are stored in config/config.php removing them from the code in dbconnect.php. This allows for easier management and editing. An install script has also been created. The install page asks for the credentials, hostname, and database name from the user and creates the tables and Admin user. The install script then writes the config.php file storing the database info.