Temporary fix for attribution of deleted users
When a user is deleted their old logs remain attributed to them. When displaying the logs they would say undefined for logs attributed to a user that didn't exist. For now a message will say Unknown User in place of the true user. I will work on reattribution of deleted users' logs.