Fixed bug where theme wasn't loading; added close button
The function getTheme was called but since dbManage.php wasn't included before the call, the page errored. I added an include_once statement (the presence script also includes dbmanage, but likewise only once). I also added a close button in the same style as the minimize button on the main page for Cheesto.