Tweaks to Cheesto mini and windowed version
The entire div containing the status options, table, and popout
button is generated in the presence.js file. The API now returns a
list of available status options. That is now used instead of a
hard coded select element in the HTML page. Styles were tweaked to
fit the new design.