Changed configuration setup
The database configuration is still in an array so it can be
injected into the dbManage class at startup. The application
settings are now directly defined as constants instead of in
one large array. Settings such as the theme directory, and app
version are still defined in the bootstrap file.