Bug fix: PHP session garbage collection never runs
Ubuntu and Debian run the PHP GC as a cron job instead of the default
way. So, the garbage collection code in session_manager.php has actually
never worked. I just assumed it worked. After some use in prod, I
discovered
that it never did. A session close now triggers garbage collection.
Eventually I may add a randomness element to it, but for now it's
fine. I also changed the database scheme to use an int instead
of timestamp type for the last_accessed field. This time I did test
the GC, and I know it works. Unfortunalty, because of the database
change, this update causes a backwards compatibility break.