Implemented base update system
The update system will be based on an update.lock file located in the
root directory. Starting with version 6, the update.lock file will
be created at setup. After which, the update script will check it
to compare versions. It will then run any updates between the old
version in the file and the new version as reported by Dandelion.
This system will be fleshed out before release.An automatic update
will not be available between v5 and v6.