Added api endpoints and refactoring
Refactored update function to accept a user id via POST. This can
be used in conjunction with a user management front end to edit
other user's statuses.

Added endpoint to read a single user's status.

Added endpoint to get status texts.

Realigned status text array with front end. Available is now 0 as
it should have been from the beginning. All the others are one
less than their previous indexes.

Removed a bit of redundant information in the return data for status
reads. In particular, since a list of texts are always returned
(eventually they won't be) the front end can simple get the text
for the status at the particular index in the text array. Eventually,
the colors and symbols won't be returned with the user statuses as
those are front end look and feel not backend.