Individualized CSS needs, JS script tag generation
Each page specifies via an array what stylesheets it needs before
the include for head.php.

The Javascript tag generation has been refactored to only use keywords
for vendor libraries and Cheesto due to permissions. The any other
arguments given are normalized with a .js extention and checked for
existance. Arguments do not require a .js unless they have the same
name as a keyword which isn't recommended.