Bigger application class, better configuration, moved lib functions
Moved a lot of what the bootstrap file was doing to the Application
class. It makes more sense for some of the processes to be done there.
Configuration has been changed to return an array on include instead
of setting constants. This configuration is accessible application
wide. The Router no longer launches the controllers, it only returns
the controller class name, method, and parameters. The Application
launches the controller. Every controller will get an instance
of the Application to access its paths, configuration, etc.

Right now some legacy code is being ran such as many of the old
constants. Eventually this wills go away and more of Dandelion
is converted over to controller, view.

Much of the old /lib files are in a utilities class Utils\View.
All the pages have been updated to use this new functions instead
of the old.