Updated install script for new format
The install script now loads the base_mysql_db.sql file
instead of a php file that had loads of prepare() execute()
statements. This is much cleaner and only executes on command
instead of dozens. Better error reporting, redirects if already
installed, redirects after install, asks for the application title
and tagline, and table prefix.

This should be a lot easier to use than the current installation
instructions.