Fixes #5 Router can handle not living in root, relative paths
The router will now build the full URL then extract the internal
routing URI to match. This ensures that, so long as the web server
and configuration are correct, the route will find a match. The
navigation bar has been changed to relative links (I thought
I had done that before, maybe I dreamed it). The login script was
the only JS file that used an absolute path. The rest appear to be
using relative paths.