Bug Fix: Adding log doesn't get root level category
For some reason jQuery isn't seeing that one of the options
in the first level select is selected. It always would select the
first option being 'Select:'. What's weirder is it only happens
when creating a new log. Editing a log works fine. It's fixed, but
I would like to know why it didn't work in the first place. It should.