Better stylesheet tag generation
The stylesheets are now checked to make sure they exist before they're
added to the document. There's no more "keyword" sets. For each style
sheet that a page wants, first it will check with a .min.css extension
and then a .css extension. It first loads the stylesheet from the first
party build folder, then for the theme. jQueryUI remains the sole exception.
It uses an if statement to check if the page is trying to load it. The
path doesn't change and is in an odd location.