Completely customizable Cheesto statuses
A status is no longer tied to an ID, instead it is now stored and
manipulated as a string. The ID system added complexity and provided
no added value. This also paves the way for completely customizable
statuses for users. The list of available "pre-made" statues can
now be defined in the app configuration. It's simply an array of
statuses. The order in the array dictates the order in the UI but
the status is not tied to the index of the array as was the case
previously.