New group management, rights, permissions, rightsAPI
Implemented updated group management system. No longer are the
permissions in the JS, everything is generated server-side and
the JS just gets all the checkboxes generated and returns the
data.

The authorized() method in Rights has been refactored to take a
variable number of arguments. This replaced the need to encapsulate
the arguments as an array. Like before if at least one permission is
true, the user has permission to perform the task.

In the Permissions class, the getGroupList method have been modified
and the corresponding repos to accepts a group name or id number.
The method will also return an extra index called permissionNames
that contains readable names for each permission indexed by the
permission name.

In the rightsAPI, the create endpoint now takes the rights parameter
as an optional parameter. The group will default to completely false
if omitted.