Added query builder tool, category module takes domid
A tool has been added to the dashboard that will help generate a search
query for the user. It doesn't provide all the available search options,
but it provides the basic and most used ones. Any thing more advanced
can be done manually.

The category TS module now takes an ID string to tell it where to render
the select boxes. Using this method the category system can only be used
by one element on a page at a time. Perhaps later this can be transformed
into a jQuery plugin for more contained usage. Also added typing to the
category module.