Fixed redirect loop on incremented version number
For now the updater only needs to update the lock file and
be happy. The updater at this point is only to have it in place
for subsequent versions. Version 6.1.0 will probably be the first
to actually utilize it. At which point it will be completely fleshed
out. It's basic right now because it's not needed yet.