Dev directories/files are an array, script help, build path
The list of directories and files to remove during build
is an array that's looped over to issues the rm command.
Makes it easier to maintain.

Using the command 'help' will print a script usage statement
offering available commands.

The build path can be given this the -p argument. It still
defaults to the current user's desktop.