Created module registry system for the api
This system is a stepping stone to plugin integration. It also
provides a clean way to map a URL path to a module method.
This also opens opportunities for later advanced matches for
a method. For example the presence of particular GET or POST parameters.
Methods already have been locked to specific HTTP methods as
deemed appropiate.