Added ability to disable users instead of delete them
Deleting a user causes issues because then logs and comments
wouldn't have a name associated with them. Disabling a user removes
it from showing up in Cheesto and doesn't allow a login, but keeps
the information for display and archival purposes. Deleting a user
is still possible, but from this point on it's not recommened.