The Users class was not a proper user class. It wasn't what I envisioned and really the others aren't either. The Users class has been replaced with a User class that is interacted with like a proper object.
A Keycard class has been implemented in the Dandelion\Auth namespace
that will help with authorization. The GateKeeper has been expanded
to handle centralized authorization. This should reduce the likelihood
of checking for one set of permissions in one place, but forgetting
one somewhere else. The authorization is based on tasks. All the current
permissions have been mapped to their correct permission. Although
most are similar, this allows for more freedom in the permissions
implementation.

A few other fixes have been added to this commit.