Added URL parameter validation to ApiCommander
When registering modules, the expected url parameters can be
given with their defaults, type, and priority (optional or required).
A call will fail if parameter validation fails. The new Validator
class acts as a wrapper around filter_var and a couple other functions
to validate most scalar types as well as objects and other filters
that PHP 5.4 supports. Validator may be extracted into a separate
library, but for now it's in a separate namespace.

The Cheesto module has been converted to the new parameter validation.

ApiException can take an HTTP status code has its 3rd constructor
argument. Defaults to 200.