Added flag to force a password reset for user
When an administrator resets a user's password, they can now set
a flag to force the user to do a password reset themselves at next
login. This way an admin can set a temp password which can be redone
by the user. This flag is available at user creation and admin reset.

For the API, the new parameter is 'force_reset'. When creating a
user it defaults to true. When resetting a password, it defaults
to false. In the UI the checkbox is set to true by default in both
places.

I also added a little extra protection in the User class.