Fixed the horrid routing
Before routing was done using the entire hostname which was
a pain and not very flexible. The router has been updated to
only the use the request uri. Dandelion has been updated with
a new configuration option call "subdirectory" which should
be the subdirectory path to Dandelion from the root hostname.
Eg. if Dandelion is located under "live" at "http://localhost/live"
then "subdirectory" would be set to "live". Any forward slashes
will be dealt with accordingly. The hostname is still used for
links and stylesheet/js paths so it should be updated if need be.
But doing it this way should make deployment a bit easier.