Added dandelion endpoint for instance data, getuser no longer needs uid
The endpoint `/api/dandelion` was added to provide information about
a running Dandelion instance. Right now it only has one method
`version` which as it implies returns the version of the current
installation.

The getuser method of the user api module no longer requires a userid.
Instead it can also use the provided api key and return information
about the "logged in" user. This will be needed for anything that
would require a user's ID or username. In particular an issue
has come up on the Cheesto extension which will require this functionality.