Expanded dev tips in CONTRIBUTING, added database to Docker Compose
The CONTRIBUTING file now has a list of dev requirements and how
to get started developing on Dandelion. It also has a new section
for developing with Docker.

Added a mariadb container to docker-compose. I hadn't added one
because there was one running on my local machine and I didn't need
it. But I felt it would be best to include the full stack. Also
updated the readme to indicate a PHP version of 5.6+ for minimum
requirements.