Fixed error page when no database is available
When attempting to show the error page, a database request
was being attempted to get a User object. Obviously this fails
when the database isn't available and so the error page itself
errors causing a somewhat cryptic error message or worse nothing
when debug is turned off. Now, the PageController constuctor
can take an optional second argument which defaults to true and
controls whether a User object is created. For the error page, it's
set to false.