Fixes #45 - Don't error when saving setting with same value
The database class was returning 0, a falsy value, since no rows
were being updated. This caused the system to think a failure had
occured. UserSettings now checks if the setting already has the
desired value and doesn't perform an update if it is.

Also changed PDO to error mode Exception and added and internal
message to APIException. Existing APIException usages will continue
working as usual. If an internal message is given, it will returned.
Otherwise the "normal" message will be returned.