Updated dependencies, PHP 7.0, removed password-compat
Composer dependencies have been updated according to the composer
file.

PHP 7.0 is now the minimum version to run. The null coalesce operator
is already in use. Other PHP 7 features will start to be used as well.

Because of the new minimum version, password-compat is no longer needed
as all PHP versions 5.5+ have the password_* functions.