Convert dashboard styles to Scss, integrate sass-loader
Dashboard and supporting styles are converted to Sass.
Less is going out of fashion and for now Sass seems like
a better bet. The dashboard TypeScript file imports the
main scss file which is compiled and saved as a separate
file in the public/dist/styles folder. The concept of
"themes" is going away and will be completely gone once
everything is in scss.

Fixed a few inconsistencies between prod and dev webpack
builds. Modified the PHP view functions to load vendored
plugins and the new styles from the proper directories.